AB upgrade app part

+++ src/enclave/utils/EJson.h
@@ -263,7 +263,7 @@ public:
     JSON(T b, typename enable_if<is_same<T, bool>::value>::type * = 0) : Internal(b), Type(Class::Boolean) {}
 
     template <typename T>
-    JSON(T i, typename enable_if<is_integral<T>::value && !is_same<T, bool>::value>::type * = 0) : Internal((long)i), Type(Class::Integral) {}
+    JSON(T i, typename enable_if<is_integral<T>::value && !is_same<T, bool>::value>::type * = 0) : Internal((long long)i), Type(Class::Integral) {}
 
     template <typename T>
     JSON(T f, typename enable_if<is_floating_point<T>::value>::type * = 0) : Internal((double)f), Type(Class::Floating) {}
@@ -309,7 +309,7 @@ public:
     typename enable_if<is_integral<T>::value && !is_same<T, bool>::value, JSON &>::type operator=(T i)
     {
         SetType(Class::Integral);
-        Internal.Int = i;
+        Internal.Int = (long long)i;
         return *this;
     }
